fix(workspaces): run setup script from TUI workspace creation paths
The initial implementation only hooked into the web API handlers, but
workspace creation in the TUI goes through src/ui/app.rs directly,
bypassing those handlers entirely. Move run_workspace_setup_script to
src/util/workspace_setup so it can be shared, and call it from all four
creation paths: web auto-create, web fork, TUI create, and TUI fork.
